+++ .gitignore
@@ -1,3 +1,5 @@
+dist/
+*.egg-info/
 *.pyc
 /__pycache__/
 /venv/
